Move pcode coverage tests to separate file
