+++ tests/main.rs
@@ -1,4 +1,5 @@
 mod common;
 mod hello_world;
+mod pcode_coverage;
 mod sleigh;
 mod x86_64_emulator;
